FE-1218 - rename file
