--- src/pages/domains/helpers/react-table-filters.js
@@ -1,14 +1,0 @@
-export function getReactTableFilters(params) {
-  return Object.entries(params)
-    .map(x => {
-      if (!x[1]) {
-        return null;
-      }
-
-      return {
-        id: x[0],
-        value: x[1],
-      };
-    })
-    .filter(Boolean);
-}
